Rename "base-detailed-schema.types.ts" to "primitives.ts"
